gold_tier/requirements.txt
# Gold Tier Autonomous Employee - Dependencies
# Python 3.11+

# Core
anthropic>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async / HTTP
httpx>=0.27.0
aiohttp>=3.9.0
aiofiles>=23.2.0
requests>=2.31.0

# Social Media APIs
tweepy>=4.14.0           # X (Twitter) API v2
facebook-sdk>=3.1.0      # Facebook / Instagram Graph API (optional)

# Scheduling
apscheduler>=3.10.0
schedule>=1.2.0

# Data / Audit
pandas>=2.1.0
openpyxl>=3.1.0          # Excel export
aiosqlite>=0.19.0        # Async SQLite
sqlalchemy>=2.0.0

# Logging
structlog>=24.1.0
rich>=13.7.0             # Pretty console output

# MCP / Tooling
mcp>=1.0.0               # Model Context Protocol SDK
fastapi>=0.111.0         # MCP HTTP transport option
uvicorn>=0.29.0

# Retry / Circuit breaker
tenacity>=8.2.0
pybreaker>=1.1.0

# Utilities
python-dateutil>=2.9.0
arrow>=1.3.0             # Friendly datetime
jinja2>=3.1.0            # CEO briefing templates
markdown>=3.6
pyyaml>=6.0.1
watchdog>=3.0.0
